+++ examples/shared_samples/enclosure.rb
@@ -26,6 +26,14 @@ require_relative '../_client' # Gives access to @client
 # api_version = 500 & variant = Synergy to OneviewSDK::API500::Synergy::Enclosure
 # api_version = 600 & variant = C7000 to OneviewSDK::API600::C7000::Enclosure
 # api_version = 600 & variant = Synergy to OneviewSDK::API600::Synergy::Enclosure
+# api_version = 800 & variant = C7000 to OneviewSDK::API800::C7000::Enclosure
+# api_version = 800 & variant = Synergy to OneviewSDK::API800::Synergy::Enclosure
+# api_version = 1000 & variant = C7000 to OneviewSDK::API1000::C7000::Enclosure
+# api_version = 1000 & variant = Synergy to OneviewSDK::API1000::Synergy::Enclosure
+# api_version = 1200 & variant = C7000 to OneviewSDK::API1200::C7000::Enclosure
+# api_version = 1200 & variant = Synergy to OneviewSDK::API1200::Synergy::Enclosure
+# api_version = 1600 & variant = C7000 to OneviewSDK::API1600::C7000::Enclosure
+# api_version = 1600 & variant = Synergy to OneviewSDK::API1600::Synergy::Enclosure
 
 # Resource Class used in this sample
 enclosure_class = OneviewSDK.resource_named('Enclosure', @client.api_version)