+++ lib/oneview-sdk/resource/api1600/synergy/enclosure_group.rb
@@ -16,20 +16,6 @@ module OneviewSDK
     module Synergy
       # Enclosure group resource implementation on API1600 Synergy
       class EnclosureGroup < OneviewSDK::API1200::Synergy::EnclosureGroup
-        # Create a resource object, associate it with a client, and set its properties.
-        # @param [OneviewSDK::Client] client The client object for the OneView appliance
-        # @param [Hash] params The options for this resource (key-value pairs)
-        # @param [Integer] api_ver The api version to use when interracting with this resource.
-        def initialize(client, params = {}, api_ver = nil)
-          @data ||= {}
-          # Default values:
-          super
-        end
-
-        def update(attributes = {})
-          @data['type'] ||= 'EnclosureGroupV8'
-          super
-        end
       end
     end
   end